requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-framebuf==1.6.6
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-rfm69==2.1.22
adafruit-circuitpython-ssd1306==2.12.18
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
autocommand==2.2.2
backports.tarfile==1.2.0
binho-host-adapter==0.1.6
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
mfrc522-python==0.0.7
more-itertools==10.3.0
packaging==24.2
pip==24.3.1
platformdirs==4.2.2
pyftdi==0.56.0
pynmea2==1.19.0
pyserial==3.5
pyusb==1.2.1
setuptools==75.6.0
tomli==2.0.1
typeguard==4.3.0
typing-extensions==4.12.2
wheel==0.38.4
zipp==3.19.2
